src/hal/lidar_interface_h.py
"""
Horizontal LIDAR Interface Module

Handles USB serial communication and data acquisition for horizontal LIDAR.
"""

import serial
import numpy as np
from typing import Optional, List, Tuple

class HorizontalLidarInterface:
    """
    Manages horizontal LIDAR (RPLIDAR) data acquisition
    
    Attributes:
        - Handles 10-15 Hz scan acquisition
        - Converts scan packets to polar ranges
        - Provides timestamped scan data
    """
    
    def __init__(self, port: str = '/dev/ttyUSB0', baudrate: int = 115200):
        """
        Initialize horizontal LIDAR interface
        
        Args:
            port: Serial port for LIDAR connection
            baudrate: Communication speed
        """
        try:
            self.serial_connection = serial.Serial(
                port=port, 
                baudrate=baudrate, 
                timeout=1.0
            )
            self._initialize_lidar()
        except serial.SerialException as e:
            raise RuntimeError(f"Failed to connect to horizontal LIDAR: {e}")
    
    def _initialize_lidar(self):
        """
        Perform LIDAR initialization sequence
        
        - Reset device
        - Configure scan mode
        - Validate connection
        """
        # TODO: Implement LIDAR initialization protocol
        pass
    
    def grab_scan(self) -> Optional[Tuple[np.ndarray, np.ndarray]]:
        """
        Acquire a single LIDAR scan
        
        Returns:
            Tuple of (angles, ranges) or None if scan fails
            - angles: numpy array of scan angles (radians)
            - ranges: numpy array of corresponding distances (meters)
        """
        try:
            # TODO: Implement scan data acquisition and parsing
            # Placeholder implementation
            angles = np.linspace(0, 2*np.pi, 360)
            ranges = np.random.uniform(0.1, 10, 360)  # Simulated ranges
            return angles, ranges
        except Exception as e:
            print(f"LIDAR scan error: {e}")
            return None
    
    def __del__(self):
        """
        Close serial connection on object destruction
        """
        if hasattr(self, 'serial_connection'):
            self.serial_connection.close()
