requirements.txt
# Core Scientific Computing
numpy==1.24.3
opencv-python==4.7.0.72
scipy==1.10.1

# Sensor Interfaces
pyserial==3.5  # For LIDAR and other serial devices
smbus2==0.4.2  # I2C interface for IMU and TOF

# Logging and Telemetry
python-json-logger==2.0.7
zmq==0.0.0  # For pub/sub communication

# Optional: ROS2 Bridge (future)
# rclpy==0.14.1

# Development and Testing
pytest==7.3.1
mypy==1.3.0
black==23.3.0

# Performance Profiling
memory-profiler==0.60.0
line-profiler==4.0.3
